Organize Delivery.format and align to spec
