+++ __TESTS__/unit/actions/Delivery.test.ts
@@ -1,10 +1,10 @@
 import {CloudinaryImage} from "../../../src/assets/CloudinaryImage";
 import CloudinaryConfig from "../../../src/config/CloudinaryConfig";
-import * as Dpr from "../../../src/values/dpr";
-import * as Quality from "../../../src/values/quality";
-import * as ColorSpace from "../../../src/values/colorSpace";
-import * as Delivery from "../../../src/actions/delivery";
 import {Format} from "../../../src/values/format";
+import {Quality} from "../../../src/values/quality";
+import {ColorSpace} from "../../../src/values/colorSpace";
+import {Dpr} from "../../../src/values/dpr";
+import {Delivery} from "../../../src/actions/delivery";
