+++ src/actions/delivery/DeliveryQuality.ts
@@ -1,12 +1,13 @@
 import {QualifierValue} from "../../internal/qualifier/QualifierValue";
-import DeliveryAction from "./DeliveryAction";
 import {Qualifier} from "../../internal/qualifier/Qualifier";
+import {DeliveryAction} from "./DeliveryAction";
 
 /**
  * @description Controls the quality of the delivered image or video.
- * @augments DeliveryAction
+ * @memberOf Actions.Delivery
+ * @extends {Actions.Delivery.DeliveryAction}
  */
-class QualityAction extends DeliveryAction {
+class DeliveryQualityAction extends DeliveryAction {
   private qualityValue: string | number;
   constructor(qualityValue: string | number) {
     super( 'q', qualityValue.toString());
@@ -30,18 +31,4 @@ class QualityAction extends DeliveryAction {
   }
 }
 
-/**
- * @description Controls the quality of the delivered image or video.
- *
- * <b>Learn more:</b> {@link https://cloudinary.com/documentation/image_optimization#how_to_optimize_image_quality | Image quality}
- *  {@link https://cloudinary.com/documentation/video_manipulation_and_delivery#quality_control | Video quality}
- * @memberOf Actions.Delivery
- * @param {string | number} qualityType For a list of supported quality types see {@link Values.qualityTypeValues| quality types} for
- * possible values.
- * @return {DeliveryAction}
- */
-function quality(qualityType:string | number) :QualityAction {
-  return new QualityAction(qualityType);
-}
-
-export default quality;
+export {DeliveryQualityAction};
